requirements.txt
python-binance
google-genai
streamlit
pandas
python-dotenv
plotly
websockets==13.1
